feng_copilot_2025_0.3/pyproject.toml
[project]
name = "nova-pasta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "snowflake-sqlalchemy>=1.7.3",
    "sqlalchemy>=2.0.38",
    "jupyter>=1.1.1",
    "notebook>=7.4.0",
    "openai>=1.14.0",
    "streamlit>=1.44.1",
    "pyarrow==19.0.0"
]
